Guess GJS path if pkg-config file not found
In order not to depend on libgjs-dev, which is annoying to install on
Travis CI, this slightly changes the check for the GJS binary's path.
Previously we used the value from GJS's pkg-config file. Now we still
prefer that, but if it is not found then we check for a binary named
"gjs". Only if that also does not exist do we fail configure.