Reporters take timer factory functions
Instead of passing in a timer object to a reporter when creating it,
pass in a function that creates timer objects. This is necessary
because with the JUnit reporter, we can have two reporters operating
at the same time, and otherwise they would share a timer object.

This will also be necessary for reporters to have multiple timers to
time suites and individual specs.