Add "store" action to option parser
Like Python's argparse, the "store" action will store the following
argument from the command line as this argument's value. If nargs == '?'
then the following argument is optional and the value from const will be
stored instead if no value is given.

The options module needs to be refactored to be more general and more
testable, but this commit adds tests as far as possible to lower the
potential barrier to writing more tests in the future. In order to test
the "store" action, this also adds the --junit option.