Stop timers once elapsed() is called
Since timers are now associated with a particular action, we don't want
to keep returning increasing values each time elapsed() is called on a
timer with a particular ID. For example, calling elapsedTime('spec:foo')
should always report the amount of time that the spec "foo" took.

This also stops the main timer in the reporter base class as soon as
jasmineDone() is called.