Return value of Command.run() is exit code
Instead of exiting directly from Command.run(), return an exit code and
let the launcher script exit -- where possible. This is in order to make
Command.run() more testable.

This is not always possible, because sometimes we have to exit from
within a callback function. In that case, we still have to call
System.exit().