Update Jasmine core to 2.2.0
This updates the internally stored copy of Jasmine to version 2.2.0. We
can get rid of a few of our local modifications, because some of them
were accepted by upstream Jasmine. Others, such as reordering functions
so as to avoid the SpiderMonkey warning about code occurring after an
unconditional return, will not be accepted, so we do need to keep a local
modifications file around.