Check for unused variables
Add the 'unused' option to our JSHint configuration and fix all the
warnings that it raises. Several times an import was not necessary any
longer but it was still left at the top of the file.

In some other cases we have to use void or array elisions to avoid the
warning. This is not so elegant so I've opened up an issue with JSHint to
see if an easier way can be found:
https://github.com/jshint/jshint/issues/2729