Split out config functions into own module
This is the start of a refactor that will separate the Jasmine binary
from an internal launcher program that will actually run the tests under
the requested environment without reading the config file twice.

We split the config-related functions from command.js out into config.js
and add ensureArray() to utils.js since it is used from both modules.

We take the opportunity to add tests to functions that didn't previously
have any test coverage.