Load config file inside Config module
This moves code from the launcher script into the Config module, and so
it becomes more testable.

At the same time, this moves some usage of the Config module out of the
Command module. In particular, this solves the problem of loading the
config file twice when setting environment variables in the config file.
(Instead, the command line options are parsed twice for now.)