Gracefully handle syntax error when importing spec file
Avoid bailing out when using the importer to read in a spec file, if it
has a syntax error. Instead, add a suite and an always-failing spec that
highlights the location of the syntax error. (Previously, an unhelpful
error message would be printed, showing the stack trace of where the
spec was imported, not the location of the syntax error within the
spec.)

Closes: #13