Gracefully handle when spec file not found
Previously, a nonexistent spec file was ignored, but this can mask a bug
in the build system: you could be passing the wrong paths and not
realize it because no tests failed. Instead, add a suite and an
always-failing spec if the spec file can't be imported due to it not
existing.

Closes: #26