Pass include paths to the subprocess using GJS_PATH instead of -I
The include paths were being interpreted as spec paths because the -I
was in the wrong place on the command line. In order to prevent this
confusion, don't use the -I argument at all (it is weird, and is
intercepted by GJS before arg parsing) and instead pass include paths to
the subprocess using the GJS_PATH environment variable.

Closes: #37