build: Substitute uninstalled variables when in subproject
If a project includes jasmine-gjs as subproject for its test
suite, it currently needs to pass a couple of environment
variables for jasmine to work uninstalled.

Now that we no longer install anything when used as subproject,
we can instead use the right substitutions, so projects can use
jasmine in their test suite without special-casing the subproject
case.