bin: Fix exit code propagation
Now that we're loading entrypoints as modules, exit codes
no longer propagate  in the same way. Therefore, we  must
do that explicitly through the system exit function.

Without this the exit code is always  ZERO and in practical
terms all CI pipelines depending on this are always passing
even when the tests are actually failing.

See 7825a0c.